+++ lib/nylas/handler/http_client.rb
