+++ lib/nylas/resources/events.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "base_resource"
-require_relative "../operations/api_operations"
+require_relative "../handler/api_operations"
 # require other mixins as needed
 
 module Nylas
